Finally making some predictions!
Need to set up the proper error measurement
using the predicted spread vs "Vegas Line".

I want to measure both the acutual regression error
with root mean square, and also the classifcation error
of the spread against the vegas line (which is what
really matters for betting).

Then it'd be nice to measure the performance of the
technique week by week over the previous 4 or 5 years,
had I made predictions at those times.